Renamed general database file.
