+++ bin/server.js
@@ -1,5 +1,5 @@
 const app = require('../app');
-const db = require('../model/contacts-db');
+const db = require('../model/db');
 require('dotenv').config();
 
 const PORT = process.env.PORT || 3000;
